initial import
This is my public demo of how to clawback from a non-TDD, non-tested,
non-releaseable piece of self-ware, and to turn it around and be nice and
Agile.  However as until 5 minutes ago this repo contained my own personal finance files and bank statements I decided that releaseing that in git history
would not be a good idea... So this is a clean .git repo, created now
after twenty minutes moving things around and then ing the old .git.